styles for tablet

+++ frontend/src/components/common/MyBouquet/MyBouquet.jsx
@@ -9,7 +9,7 @@ import MyBouquetEmpty from './components/MyBuoquetEmpty';
 import { useWindowSize } from '../../../hooks/useWindowSize';
 import { useSelector } from 'react-redux/es/hooks/useSelector';
 import { useNavigate } from 'react-router-dom';
-
+import Footer from '../../Footer';
 import { ICONS } from '../../ui-kit/icons';
 import styles from './MyBouquet.module.scss';
 import { useDispatch } from 'react-redux';
@@ -80,7 +80,7 @@ const MyBouquet = ({ toggleMyBouquet }) => {
       <div className={styles.container}>
         <div className={styles.headerBlock}>
           <p className={styles.title}>
-            {width > 767 ? 'Створення власного букету' : 'Створення букету'}
+            {width > 867 ? 'Створення власного букету' : 'Створення букету'}
           </p>
           <IconButton
             icon={<ICONS.CloseIcon className={styles.icon} />}
@@ -90,33 +90,29 @@ const MyBouquet = ({ toggleMyBouquet }) => {
         {hasItems ? <MyBouquetItem items={bouquetItems} /> : <MyBouquetEmpty />}
       </div>
       <div className={styles.bottomBlock}>
-        <div className={hasItems ? styles.reverse : ''}>
-          <div className={styles.changeItems}>
-            <span
-              className={hasItems ? styles.deleteAll : styles.noActiveDelete}
-              onClick={handleDeleteAll}
-              disabled={!hasItems}
-            >
-              Видалити все
-            </span>
-            {isModalOpen && (
-              <ConfirmationPopup
-                setIsOpen={setModalOpen}
-                deleteAll={clearAll}
-              />
-            )}
-            <Button
-              label={width > 767 ? 'Додати компонент' : 'Додати'}
-              variant="secondary"
-              padding="padding-sm"
-              icon={<ICONS.addComponent />}
-              onClick={addMore}
-            />
-          </div>
-          <div className={hasItems ? styles.divider : ''}>
-            <Divider />
-          </div>
+        <div className={styles.changeItems}>
+          <span
+            className={hasItems ? styles.deleteAll : styles.noActiveDelete}
+            onClick={handleDeleteAll}
+            disabled={!hasItems}
+          >
+            Видалити все
+          </span>
+          {isModalOpen && (
+            <ConfirmationPopup setIsOpen={setModalOpen} deleteAll={clearAll} />
+          )}
+          <Button
+            label={width > 767 ? 'Додати компонент' : 'Додати'}
+            variant="secondary"
+            padding="padding-sm"
+            icon={<ICONS.addComponent />}
+            onClick={addMore}
+          />
         </div>
+        <div className={styles.divider}>
+          <Divider />
+        </div>
+
         <div className={styles.bottom}>
           <div className={styles.total}>
             <p>Разом:</p>
@@ -127,15 +123,18 @@ const MyBouquet = ({ toggleMyBouquet }) => {
               <span>грн</span>
             </div>
           </div>
-          <Button
-            label="Зібрати букет"
-            variant={hasItems ? 'primary' : 'disabled'}
-            padding="padding-even"
-            onClick={addBouquetToCart}
-            disabled={bouquetItems.length === 0}
-          />
+          <div className={styles.button}>
+            <Button
+              label="Зібрати букет"
+              variant={hasItems ? 'primary' : 'disabled'}
+              padding="padding-even"
+              onClick={addBouquetToCart}
+              disabled={bouquetItems.length === 0}
+            />
+          </div>
         </div>
       </div>
+      {width < 868 && <Footer />}
     </Modals>
   );
 };
